Break main method out of Universe
